GO TO BEEEEEEEED

+++ ai.py
@@ -1,6 +1,6 @@
 import replicate
 import requests
-
+from os.path import exists
 
 rep = replicate.Client(api_token="e53c3e92e92cedf6f1e24a377e261efa92d83a0d")
 model = rep.models.get("stability-ai/stable-diffusion")
@@ -9,6 +9,7 @@ model = rep.models.get("stability-ai/stable-diffusion")
 def generate_image(image_state):
     prompt = image_state.content[0]
     file_name = image_state.id
-    data = requests.get(model.predict(prompt=prompt)[0]).content
-    with open(f"images/{file_name}.png", "wb") as file:
-        file.write(data)
+    if not exists(f'images/{file_name}.png'):
+        data = requests.get(model.predict(prompt=prompt)[0]).content
+        with open(f"images/{file_name}.png", "wb") as file:
+            file.write(data)
